Move data models out into a separate file
