+++ backend/api/handlers.go
@@ -96,10 +96,6 @@ func fetchIssueSubjects(c *fiber.Ctx, entries []Entry) (bool, error) {
 	return true, nil
 }
 
-type user struct {
-	Login string `json:"login"`
-}
-
 // loginHandler godoc
 // @Summary	Log in a user
 // @Description	Log in a user using the Redmine API
@@ -183,28 +179,6 @@ func logoutHandler(c *fiber.Ctx) error {
 	}
 
 	return c.SendStatus(fiber.StatusNoContent)
-}
-
-type Issue struct {
-	Id      int    `json:"id"`
-	Subject string `json:"subject"`
-}
-
-type Activity struct {
-	Id   int    `json:"id"`
-	Name string `json:"name"`
-}
-
-type Entry struct {
-	Issue    Issue    `json:"issue"`
-	Activity Activity `json:"activity"`
-}
-
-type PriorityEntry struct {
-	Issue      Issue    `json:"issue"`
-	Activity   Activity `json:"activity"`
-	CustomName string   `json:"custom_name"`
-	IsHidden   bool     `json:"is_hidden"`
 }
 
 // recentIssuesHandler godoc
